requirements.txt
dash
pandas
numpy
plotly
scikit-learn
xgboost
lightgbm
imbalanced-learn
networkx
torch
torch-geometric
fastapi
uvicorn
mlflow
prometheus-client
grafana-api
sentry-sdk
faker
psycopg2-binary
